Hopefully this fixes slowness
